Many changes in BasinBuilder and the classes taht are here hust for it.
BasinBuilder's algorithms are improved to be more accurate when computing the basin graph. We don't use contains but covers to check if a polygon or linestring is already in the basin graph.

EdgePart and PointPart both have a customized toString method now.

PointPart has been documented.

Minor change in ST_BasinGraph.