The pom.xml file has been cleaned.
We definitely must use transitive dependencies in maven as far as we can !!! It will avoid mismanagement of the versions of external libraries. For instance, we were depending on jts 1.11 through GDMS, and jts 1.10 through our explicit declaration in tanato. Consequently, we couldn't be sure of the coherence of the operations made in Tanato, as we're making explicit operations through geometries (through jts 1.10 until now) and implicit operations through GDMS (through jts 1.11).